styled ended hunt page for hunter view and added router link back to join hunt page.
