--- client/src/app/hunters/ended-hunt/hunter-ended-hunt/hunter-ended-hunt.component.html
@@ -1,1 +1,0 @@
-<p>hunter-ended-hunt works!</p>
