--- client/src/app/hunters/ended-hunt/hunter-ended-hunt/hunter-ended-hunt.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-hunter-ended-hunt',
-  standalone: true,
-  imports: [],
-  templateUrl: './hunter-ended-hunt.component.html',
-  styleUrl: './hunter-ended-hunt.component.scss'
-})
-export class HunterEndedHuntComponent {
-
-}
