+++ client/src/app/hunters/hunter-ended-hunt/hunter-ended-hunt.component.html
@@ -1,0 +1,8 @@
+<body>
+<mat-card class="completion-card">
+  <mat-card-title class="completion-title">Hunt complete, return to host!</mat-card-title>
+  <mat-card-actions class="return-button">
+    <button mat-raised-button color="primary" [routerLink]="['/hunters/']">Join Another Hunt</button>
+  </mat-card-actions>
+</mat-card>
+</body>
